refactor(imports): move reauth-reason enum out of coordinator package
The reauth-reason model was introduced under coordinator/_reauth_reason.py,
which made api.py import from the coordinator package (api.py -> coordinator).
Since coordinator/__init__.py eagerly imports `from ..api import GoogleFindMyAPI`
(coordinator -> api), this created a layering violation and a latent import
cycle: api.py sits below the coordinator package, so it must not depend on it.

The cycle is currently masked by the package __init__ import order (diagnostics
loads coordinator before api on a cold import), but that masking is accidental
and fragile: a reordering of the diagnostics import chain would surface the
`ImportError: cannot import name 'GoogleFindMyAPI' from partially initialized
module` that Codex flagged.

Fix (Move Module refactoring): the enum is a dependency-light shared contract
(stdlib-only: dataclasses + enum) that both api.py (low) and coordinator/* (high)
need, so it belongs at the package root, below both. Moving it removes the
api -> coordinator edge entirely; api.py no longer imports from the coordinator
package. All 8 production import sites and 6 test import sites are updated
(external importers use the root path, coordinator-internal importers use `..`).

Behavior-preserving: no enum value or logic changed (git rename similarity 98%,
only the file-header path comment differs). Verified: api can be imported in
isolation without dragging coordinator/__init__; full suite 4668 passed;
ruff 0.14.14 check+format and mypy green; independent [REVIEW:DIFF] PASS.
