+++ tests/test_coordinator_semantic_mappings.py
@@ -8,12 +8,12 @@ from unittest.mock import MagicMock
 import pytest
 from homeassistant.config_entries import ConfigEntryAuthFailed
 
+from custom_components.googlefindmy._reauth_reason import ReauthReasonCode
 from custom_components.googlefindmy.const import OPT_SEMANTIC_LOCATIONS
 from custom_components.googlefindmy.coordinator import (
     CryptoStatus,
     GoogleFindMyCoordinator,
 )
-from custom_components.googlefindmy.coordinator._reauth_reason import ReauthReasonCode
 from custom_components.googlefindmy.coordinator.polling import (
     _MAX_DECRYPT_FAILURES,
     _MAX_TRANSIENT_AUTH_FAILURES,
